requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
aiosqlite==0.20.0
python-dotenv==1.0.1
Pillow==10.3.0
requests==2.32.3
